+++ src/main/java/com/github/blutorange/maven/plugin/closurecompiler/common/AggregationType.java
@@ -1,0 +1,8 @@
+package com.github.blutorange.maven.plugin.closurecompiler.common;
+
+/**
+ * Defines the aggregation type.
+ */
+public enum AggregationType {
+js
+}