+++ src/main/java/com/github/blutorange/maven/plugin/closurecompiler/plugin/ProcessJSFilesTask.java
@@ -22,6 +22,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.io.Writer;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -42,6 +43,12 @@ import com.google.javascript.jscomp.LightweightMessageFormatter;
 import com.google.javascript.jscomp.MessageFormatter;
 import com.google.javascript.jscomp.SourceFile;
 import com.google.javascript.jscomp.SourceMap;
+
+import eu.maxschuster.dataurl.DataUrl;
+import eu.maxschuster.dataurl.DataUrlBuilder;
+import eu.maxschuster.dataurl.DataUrlEncoding;
+import eu.maxschuster.dataurl.DataUrlSerializer;
+import eu.maxschuster.dataurl.IDataUrlSerializer;
 
 /**
  * Task for merging and compressing JavaScript files.
@@ -98,13 +105,15 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
   protected void minify(List<File> srcFiles, File minifiedFile) throws IOException {
     minifiedFile.getParentFile().mkdirs();
 
-    try (OutputStream out = new FileOutputStream(minifiedFile);
-        OutputStreamWriter writer = new OutputStreamWriter(out, charset)) {
+    try (OutputStream out = new FileOutputStream(minifiedFile); Writer writer = new OutputStreamWriter(out, charset)) {
+
       log.info("Creating the minified file [" + ((verbose) ? minifiedFile.getPath() : minifiedFile.getName())
           + "].");
 
       log.debug("Using Google Closure Compiler engine.");
 
+      // List<SourceFile> sourceFileList = srcFiles.stream().map(f -> SourceFile.fromPath(f.toPath(),
+      // charset)).collect(Collectors.toList());
       List<SourceFile> sourceFileList = new ArrayList<SourceFile>();
       for (File srcFile : srcFiles) {
         InputStream in = new FileInputStream(srcFile);
@@ -112,61 +121,43 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
         sourceFileList.add(input);
       }
 
+      // Set common options
       CompilerOptions options = new CompilerOptions();
-      closureConfig.getCompilationLevel().setOptionsForCompilationLevel(options);
-      options.setOutputCharset(charset);
-      options.setLanguageIn(closureConfig.getLanguageIn());
-      options.setLanguageOut(closureConfig.getLanguageOut());
-      options.setAngularPass(closureConfig.getAngularPass());
-      options.setColorizeErrorOutput(closureConfig.getColorizeErrorOutput());
-      options.setDependencyOptions(closureConfig.getDependencyOptions());
-      options.setExtraAnnotationNames(closureConfig.getExtraAnnotations());
-      options.setDefineReplacements(closureConfig.getDefineReplacements());
-      options.setSourceMapIncludeSourcesContent(closureConfig.getIncludeSourcesContent());
+      setCommonOptions(options);
 
-      File sourceMapResult = new File(minifiedFile.getPath() + ".map");
+      // Tell the compiler to create a source map, if configured.
+      File sourceMapFile = new File(minifiedFile.getPath() + ".map");
       if (closureConfig.getSourceMapFormat() != null) {
         options.setSourceMapFormat(closureConfig.getSourceMapFormat());
-        options.setSourceMapOutputPath(sourceMapResult.getPath());
-        // options.setSourceMapLocationMappings(Lists.newArrayList(new
-        // SourceMap.LocationMapping(sourceDir.getPath() + File.separator, "")));
+        options.setSourceMapIncludeSourcesContent(closureConfig.getIncludeSourcesContent());
+        options.setSourceMapOutputPath(sourceMapFile.getPath());
       }
 
+      // Set warning levels
       if (closureConfig.getWarningLevels() != null) {
         for (Map.Entry<DiagnosticGroup, CheckLevel> warningLevel : closureConfig.getWarningLevels().entrySet()) {
           options.setWarningLevel(warningLevel.getKey(), warningLevel.getValue());
         }
       }
 
-      List<SourceFile> externs = closureConfig.getExterns();
+      // Set (external) libraries to be available
+      List<SourceFile> externs = new ArrayList<>();
       externs.addAll(CommandLineRunner.getBuiltinExterns(closureConfig.getEnvironment()));
       externs.addAll(closureConfig.getExterns());
 
-      Compiler compiler = new Compiler();
+      // Now compile
+      final Compiler compiler = new Compiler();
       compiler.compile(externs, sourceFileList, options);
 
       // Check for errors.
-      JSError[] errors = compiler.getErrors();
-      if (errors.length > 0) {
-        StringBuilder msg = new StringBuilder("JSCompiler errors\n");
-        MessageFormatter formatter = new LightweightMessageFormatter(compiler);
-        for (JSError e : errors) {
-          msg.append(formatter.formatError(e));
-        }
-        throw new RuntimeException(msg.toString());
-      }
+      checkForErrors(compiler);
 
+      // Write compiled file to output file
       writer.append(compiler.toSource());
 
+      // Create source map if configured.
       if (closureConfig.getSourceMapFormat() != null) {
-        log.info("Creating the minified files map ["
-            + ((verbose) ? sourceMapResult.getPath() : sourceMapResult.getName()) + "].");
-
-        sourceMapResult.createNewFile();
-        flushSourceMap(sourceMapResult, minifiedFile.getName(), compiler.getSourceMap());
-
-        writer.append(System.getProperty("line.separator"));
-        writer.append("//# sourceMappingURL=" + sourceMapResult.getName());
+        createSourceMap(writer, compiler, minifiedFile, sourceMapFile);
       }
     }
     catch (IOException e) {
@@ -180,14 +171,73 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
     logCompressionGains(srcFiles, minifiedFile);
   }
 
-  private boolean flushSourceMap(File sourceMapOutputFile, String minifyFileName, SourceMap sourceMap) {
-    try (BufferedWriter out = Files.newBufferedWriter(sourceMapOutputFile.toPath(), StandardCharsets.UTF_8)) {
+  private void setCommonOptions(CompilerOptions options) {
+    closureConfig.getCompilationLevel().setOptionsForCompilationLevel(options);
+    options.setOutputCharset(charset);
+    options.setLanguageIn(closureConfig.getLanguageIn());
+    options.setLanguageOut(closureConfig.getLanguageOut());
+    options.setAngularPass(closureConfig.getAngularPass());
+    options.setColorizeErrorOutput(closureConfig.getColorizeErrorOutput());
+    options.setDependencyOptions(closureConfig.getDependencyOptions());
+    options.setExtraAnnotationNames(closureConfig.getExtraAnnotations());
+    options.setDefineReplacements(closureConfig.getDefineReplacements());
+  }
+
+  private void checkForErrors(Compiler compiler) {
+    JSError[] errors = compiler.getErrors();
+    if (errors.length > 0) {
+      StringBuilder msg = new StringBuilder("JSCompiler errors\n");
+      MessageFormatter formatter = new LightweightMessageFormatter(compiler);
+      for (JSError e : errors) {
+        msg.append(formatter.formatError(e));
+      }
+      throw new RuntimeException(msg.toString());
+    }
+  }
+
+  private void createSourceMap(Writer writer, Compiler compiler, File minifiedFile, File sourceMapFile) throws IOException {
+    log.info("Creating the minified files map ["
+        + ((verbose) ? sourceMapFile.getPath() : sourceMapFile.getName()) + "].");
+
+    switch (closureConfig.getSourceMapOutputType()) {
+      case inline:
+        StringBuilder sb = new StringBuilder();
+        compiler.getSourceMap().appendTo(sb, minifiedFile.getName());
+        DataUrl unserialized = new DataUrlBuilder() //
+            .setMimeType("application/json") //
+            .setEncoding(DataUrlEncoding.BASE64) //
+            .setData(sb.toString().getBytes(StandardCharsets.UTF_8)) //
+            .setHeader("charset", "utf-8") //
+            .build();
+        IDataUrlSerializer serializer = new DataUrlSerializer();
+        String dataUrl = serializer.serialize(unserialized);
+        writer.append(System.getProperty("line.separator"));
+        writer.append("//# sourceMappingURL=" + dataUrl);
+        break;
+      case file:
+        sourceMapFile.createNewFile();
+        flushSourceMap(sourceMapFile, minifiedFile.getName(), compiler.getSourceMap());
+        break;
+      case reference:
+        sourceMapFile.createNewFile();
+        flushSourceMap(sourceMapFile, minifiedFile.getName(), compiler.getSourceMap());
+        writer.append(System.getProperty("line.separator"));
+        writer.append("//# sourceMappingURL=" + sourceMapFile.getName());
+        break;
+      default:
+        log.warn("unknown source map inclusion type: " + closureConfig.getSourceMapOutputType());
+        throw new RuntimeException("unknown source map inclusion type: " + closureConfig.getSourceMapOutputType());
+    }
+  }
+
+  private boolean flushSourceMap(File sourceMapFile, String minifyFileName, SourceMap sourceMap) {
+    try (BufferedWriter out = Files.newBufferedWriter(sourceMapFile.toPath(), StandardCharsets.UTF_8)) {
       sourceMap.appendTo(out, minifyFileName);
       return true;
     }
     catch (IOException e) {
       log.error("Failed to write the JavaScript Source Map file ["
-          + (verbose ? sourceMapOutputFile.getPath() : sourceMapOutputFile.getName()) + "].", e);
+          + (verbose ? sourceMapFile.getPath() : sourceMapFile.getName()) + "].", e);
       return false;
     }
   }